+++ vector/src/main/java/im/vector/app/features/media/DataAttachmentRoomProvider.kt
@@ -77,7 +77,6 @@ class DataAttachmentRoomProvider(
     override fun getFileForSharing(position: Int, callback: (File?) -> Unit) {
         val item = getItem(position)
         fileService.downloadFile(
-                downloadMode = FileService.DownloadMode.FOR_EXTERNAL_SHARE,
                 id = item.eventId,
                 fileName = item.filename,
                 mimeType = item.mimeType,